Wiring 6 legs with idenpendent IK
ie each leg is running its own IK calculations, instead of the previous
version where IK was solved for one leg and copied to all others.
Visually the robot does the same things, but this proves that the
hardware is good enough to run my lazy code in "real time" for all 18
servos.